requirements.txt
aiogram==3.13.0
asyncpg==0.30.0
google-generativeai==0.8.3
Pillow==11.1.0
python-dotenv==1.0.1
